+++ .storybook/main.ts
@@ -32,8 +32,10 @@ const config: StorybookConfig = {
       button[id="removed-reveal"] {display: none; }
       .sidebar-item:has(button[id="buttons-button-rive"]) { display: none; }
       .sidebar-item:has(button[id="buttons-button-split"]) { display: none; }
+      
       div[id="wip"] > button:first-child::after { content: " 🚧 "; }
       button[id^="wip-"]::after { content: " 🚧 "; }
+      a[id$="-wip"]::after { content: " 🚧 "; }
     </style>
   `,