+++ src/components/Paper/Paper.tsx
@@ -4,8 +4,9 @@ import { Color } from '../../utils/colors';
 
 export type PaperProps = {
   children: React.ReactNode;
-  element?: 'section' | 'article' | 'main' | 'aside' | 'div';
-  elevation?: -1 | 0 | 1 | 2 | 3;
+  element?: 'section' | 'article' | 'main' | 'aside' | 'div' | 'dialog';
+  id?: string;
+  elevation?: -2 | 0 | 2 | 4 | 6;
   roundedCorner?: boolean;
   backgroundColor?: Color;
   className?: string;
@@ -17,6 +18,7 @@ export type PaperProps = {
 export const Paper: React.FC<PaperProps> = ({
   children,
   element: Component = 'div',
+  id,
   elevation,
   roundedCorner = false,
   backgroundColor = '--connect__white',
@@ -43,7 +45,7 @@ export const Paper: React.FC<PaperProps> = ({
   }
 
   return (
-    <Component className={paperClassName} data-testid={dataTestId} style={paperProps}>
+    <Component id={id} className={paperClassName} data-testid={dataTestId} style={paperProps}>
       {children}
     </Component>
   );